phase4/requirements.txt
torch>=1.9.0
numpy>=1.19.2
gymnasium>=0.28.1
minari>=0.3.0
d4rl @ git+https://github.com/rail-berkeley/d4rl@master#egg=d4rl
wandb>=0.12.0
pyyaml>=5.4.1
tqdm>=4.62.3
matplotlib>=3.4.3 